Move components to web-components/
