+++ src/gatsby-shell/comps/SheetTemplateView.tsx
@@ -4,14 +4,14 @@ import { useSheetContext } from '../SheetTemplate'
 
 // Components
 import CommentsArea from '../../web-comments/CommentsArea'
+import IntroContent from '../../web-components/IntroContent'
+import MainHeading from '../../web-components/MainHeading'
+import MiniMarkdown from '../../web-components/MiniMarkdown'
+import PreFooter from '../../web-components/PreFooter'
+import RelatedPostsArea from '../../web-components/RelatedPostsArea'
+import SearchFooter from '../../web-components/SearchFooter'
+import TopNav from '../../web-components/TopNav'
 import PostContent from '../../web-post-content/PostContent'
-import IntroContent from '../../web/components/IntroContent'
-import MainHeading from '../../web/components/MainHeading'
-import MiniMarkdown from '../../web/components/MiniMarkdown'
-import PreFooter from '../../web/components/PreFooter'
-import RelatedPostsArea from '../../web/components/RelatedPostsArea'
-import SearchFooter from '../../web/components/SearchFooter'
-import TopNav from '../../web/components/TopNav'
 import MetaTags from '../MetaTags'
 import CommonHead from './CommonHead'
