+++ src/web-components/HomeButton.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'gatsby'
 import React from 'react'
-import { homeLine } from '../../web-icons'
+import { homeLine } from '../web-icons'
 import CSS from './HomeButton.module.css'
 
 /**
