+++ src/web-components/IntroContent.tsx
