+++ src/web-components/RelatedPostsCallout.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'gatsby'
 import React from 'react'
-import useSiteContent from '../../gatsby-hooks/useSiteContent'
+import useSiteContent from '../gatsby-hooks/useSiteContent'
 import CSS from './RelatedPostsCallout.module.css'
 import SpanPushButton from './SpanPushButton'
