+++ src/web-components/RelatedPostsSection.tsx
