+++ src/web-components/SearchFooter.tsx
